Improve memory, stream safety and profiling checks
- Use safer profiling macro checks in geMemoryAllocator.h
- Enhance type safety and bounds checking in MemoryDataStream methods
- Fix object slicing in clone method with static_cast
- Use correct types for stream offsets and buffer sizes
- Update AppVeyor badge to SVG in README
- Clean up comments and remove redundant casts in gePosixCrashHandler.cpp

+++ sdk/geUtilities/src/posix/gePosixCrashHandler.cpp
@@ -10,13 +10,13 @@
  *
  * @bug	    No known bugs.
  */
- /*****************************************************************************/
+/*****************************************************************************/
 #ifndef _WIN32
 /*****************************************************************************/
 /**
  * Includes
  */
- /*****************************************************************************/
+/*****************************************************************************/
 #include "gePrerequisitesUtilities.h"
 #include "geCrashHandler.h"
 #include "geDebug.h"
@@ -109,7 +109,7 @@ namespace geEngineSDK {
   CrashHandler::getStackTrace() {
 #if defined(__unix__) || defined(__APPLE__)
     void* addrs[GE_MAX_STACKTRACE_DEPTH];
-    const int n = ::backtrace(addrs, static_cast<int>(GE_MAX_STACKTRACE_DEPTH));
+    const int n = ::backtrace(addrs, GE_MAX_STACKTRACE_DEPTH);
 
     StringStream ss;
     ss << "Stack frames: " << n << "\n";
